Laravel 11.x Shift (#561)
Update to Laravel 11

---------

+++ app/Modules/QueueMonitor/src/Listeners/JobFailedListener.php
@@ -21,6 +21,10 @@ class JobFailedListener
 
         $jobStartTime = Cache::pull($key);
 
-        Log::info('Job failed', ['job' => $jobName, 'duration' => now()->diffInSeconds($jobStartTime), 'exception' => $event->exception]);
+        Log::info('Job failed', [
+            'job' => $jobName,
+            'duration' => (int)now()->diffInSeconds($jobStartTime),
+            'exception' => $event->exception
+        ]);
     }
 }
